pyproject.toml
[project]
name = "conlex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "llama-index>=0.12.52",
    "loguru>=0.7.3",
    "openai>=1.97.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]

[tool.setuptools]
package-dir = {conlex = "src/conlex"}
